feature: Organize classes and web services
- Move classes into their own files with associated web services
- Created a PokemonKit class to conveniently access individual services
